+++ src/evm/releases/lockup/v1.0/abi.ts
@@ -7,4 +7,4 @@ export const abi = {
   [manifest.core.SABLIER_V2_LOCKUP]: sablierV2LockupAbi,
   [manifest.core.SABLIER_V2_LOCKUP_DYNAMIC]: sablierV2LockupDynamicAbi,
   [manifest.core.SABLIER_V2_LOCKUP_LINEAR]: sablierV2LockupLinearAbi,
-};
+} as const;